// src/shared/generated/GraphicsSettingsRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GpuRasterizationMode } from "./GpuRasterizationMode";
import type { HardwareVideoDecodeMode } from "./HardwareVideoDecodeMode";

export type GraphicsSettingsRecord = { hardwareAcceleration: boolean, rasterization: GpuRasterizationMode, videoDecode: HardwareVideoDecodeMode, };